+++ test/unit/data-migration.test.ts
@@ -355,7 +355,6 @@ config.parallel('data-migration.test.js', () => {
 
                     const undeleted = await getAllDocuments(
                         old.schema.primaryPath,
-                        old.database.storage,
                         old.storageInstance
                     );
                     const amount = undeleted.length;