+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Refactor usage of RxCollection.storageInstance to ensure all hooks run properly.
 - Added `withDeleted` parameter to `RxStorageKeyObjectInstance.findLocalDocumentsById()`
 - The `_deleted` field is now required for each data interaction with `RxStorage`.
+- `RxStorageInstance` must have the `RxStorage` in the `storage` property.
 
 - Removed the `core` plugin. The default export `from 'rxdb'` now exports only the RxDB core without plugins.
 - Added optional `init()` function to `RxPlugin`.